Changed devID: Number, to String. The Device ID for Iris devices can sometimes contain letters.
Changed config/multer.js, I removed the Date.now since the iris image
name already contains a unix timestamp. If you need this and add it back
in let me know because it will effect how I parse the check response.

I changed the /sync_date error handling, as the
res.status(123).send("error message") will not override the default
error message for a given status type, even if one hasn't been assigned.
By doing it this way the "No register found" error is sent across
correctly.

Did the same thing with the /images error handling

I changed the output of /check to allow me to do less parsing on the
iris side. However, this still results in a bit of unnecesary parsing on
the iris side. Right now this will return
/images/<timestamp>-<photoid>.jpg where in reality I only want
<timestamp>. However, for now this currently works fine.

I changed the getWanIp method again, this is mostly to work on my
machine, as I have several network interfaces. I don't know how
necessary this is for running on a tablet.

I changed the syncDate method to sort on the "Data" field of the images,
as this contains the timestamp. So far this is working well.

I also removed the Pid filter. PID stands for the patient ID not the
provider ID.

Furthermore, the current NCI study stance has been that any provider can
upload all patient data, despite not having direct access to that data.
So filtering on provider data would be a departure from how other sites
outside of brazil are operating. We should talk about this in more depth
to see if this is something Brazil wants different from the rest of the
sites.